codigo/Live177/pyproject.toml
[tool.poetry]
name = "live177"
version = "0.1.0"
description = ""
authors = ["dunossauro <mendesxeduardo@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pygubu-designer = "^0.19"
googletrans = "3.1.0a0"
ttkthemes = "^3.2.2"
ttkbootstrap = "^0.5.1"
cx-Freeze = "^6.8"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
